Move all Recipient warnings and lists to its own component
