+++ ui/components/multichain/pages/send/components/index.ts
@@ -4,3 +4,4 @@ export { SendPageNetworkPicker } from './network-picker';
 export { SendPageYourAccount } from './your-accounts';
 export { SendPageRecipientInput } from './recipient-input';
 export { SendPageAddressBook } from './address-book';
+export { SendPageRecipient } from './recipient';
